move make_VLM_wings() to its own file.
also moved its helper functions
