+++ trunk/SUAVE/Methods/Aerodynamics/Common/Fidelity_Zero/Lift/__init__.py
@@ -14,4 +14,5 @@ from .compute_wing_wake                       import compute_wing_wake
 from .compute_propeller_nonuniform_freestream import compute_propeller_nonuniform_freestream
 from .generate_vortex_distribution            import generate_vortex_distribution, compute_unit_normal 
 from .fuselage_correction                     import fuselage_correction
+from .make_VLM_wings                          import make_VLM_wings
 from .VLM                                     import VLM
