+++ trunk/SUAVE/Methods/Flight_Dynamics/Static_Stability/Approximations/Supporting_Functions/convert_sweep.py
@@ -3,6 +3,7 @@
 #
 # Created:  Feb 2014, T. Momose
 # Modified: Jan 2016, E. Botero
+#           Jun 2021, A. Blaufox
 
 # ----------------------------------------------------------------------
 #  Imports
